Renamed App to Main in preparation for using Grommet App component
